unexpose SignalPulse for now - src/algorithms is not the place for meta utilities
